services/document-service/pyproject.toml
[project]
name = "office-agent-document-service"
version = "0.1.0"
description = "Local document processing service for OfficeAgent"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
  "pypdf>=4.3.0",
  "python-docx>=1.1.2"
]

[tool.uvicorn]
host = "127.0.0.1"
port = 8765
